requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
djangorestframework>=3.15
Pillow>=10.0
